# openapi.yaml
openapi: 3.1.0
info:
  title: HHC Web API
  version: 0.2.0
servers:
  - url: https://www.alive.org.tw/api
paths:
...
  /news:
    get:
      operationId: listPublicNews
      parameters: [{ $ref: '#/components/parameters/Locale' }]
      responses: { '200': { $ref: '#/components/responses/PublicContentList' } }
  /history:
    get:
      operationId: listPublicHistory
      parameters: [{ $ref: '#/components/parameters/Locale' }]
      responses: { '200': { $ref: '#/components/responses/PublicContentList' } }
  /videos:
    get:
      operationId: listPublicVideos
      parameters: [{ $ref: '#/components/parameters/Locale' }]
      responses: { '200': { $ref: '#/components/responses/PublicContentList' } }
  /home:
    get:
      operationId: getPublicHome
      parameters: [{ $ref: '#/components/parameters/Locale' }]
      responses:
        '200':
          description: Published home page projection
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HomeEnvelope' }
...
  /admin/content/{module}:
    get:
      operationId: listAdminContent
      security: [{ bearerAuth: [] }]
      x-required-scopes: [cms:read]
      parameters:
        - $ref: '#/components/parameters/ContentModule'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - { name: status, in: query, schema: { $ref: '#/components/schemas/ContentStatus' } }
      responses:
        '200':
          description: Typed content list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContentListEnvelope' }
    post:
      operationId: createContent
      security: [{ bearerAuth: [] }]
      x-required-scopes: [cms:write]
      parameters:
        - $ref: '#/components/parameters/ContentModule'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody: { $ref: '#/components/requestBodies/ContentWriteInput' }
      responses:
        '201': { $ref: '#/components/responses/ContentItem' }
        '409': { $ref: '#/components/responses/Error' }
  /admin/content/{module}/{contentId}:
    get:
      operationId: getAdminContent
      security: [{ bearerAuth: [] }]
      x-required-scopes: [cms:read]
      parameters:
        - $ref: '#/components/parameters/ContentModule'
        - $ref: '#/components/parameters/ContentID'
      responses:
        '200': { $ref: '#/components/responses/ContentItem' }
        '404': { $ref: '#/components/responses/Error' }
    put:
      operationId: updateContent
      security: [{ bearerAuth: [] }]
      x-required-scopes: [cms:write]
      parameters:
        - $ref: '#/components/parameters/ContentModule'
        - $ref: '#/components/parameters/ContentID'
        - $ref: '#/components/parameters/IfMatch'
      requestBody: { $ref: '#/components/requestBodies/ContentWriteInput' }
      responses:
        '200': { $ref: '#/components/responses/ContentItem' }
        '412': { $ref: '#/components/responses/Error' }
  /admin/content/{module}/{contentId}/publish:
    post:
      operationId: publishContent
      security: [{ bearerAuth: [] }]
      x-required-scopes: [cms:publish]
      parameters:
        - $ref: '#/components/parameters/ContentModule'
        - $ref: '#/components/parameters/ContentID'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200': { $ref: '#/components/responses/ContentItem' }
        '422': { $ref: '#/components/responses/Error' }
  /admin/content/{module}/{contentId}/unpublish:
    post:
      operationId: unpublishContent
      security: [{ bearerAuth: [] }]
      x-required-scopes: [cms:publish]
      parameters:
        - $ref: '#/components/parameters/ContentModule'
        - $ref: '#/components/parameters/ContentID'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200': { $ref: '#/components/responses/ContentItem' }
  /admin/content/{module}/{contentId}/revisions:
    get:
      operationId: listContentRevisions
      security: [{ bearerAuth: [] }]
      x-required-scopes: [cms:read]
      parameters:
        - $ref: '#/components/parameters/ContentModule'
        - $ref: '#/components/parameters/ContentID'
      responses:
        '200':
          description: Content revisions
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContentRevisionListEnvelope' }
  /admin/content/{module}/{contentId}/revisions/{revision}/restore:
    post:
      operationId: restoreContentRevision
      security: [{ bearerAuth: [] }]
      x-required-scopes: [cms:write]
      parameters:
        - $ref: '#/components/parameters/ContentModule'
        - $ref: '#/components/parameters/ContentID'
        - $ref: '#/components/parameters/Revision'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200': { $ref: '#/components/responses/ContentItem' }
  /admin/content/news/{contentId}/upload-sessions:
    post:
      operationId: createNewsCoverUpload
      security: [{ bearerAuth: [] }]
      x-required-scopes: [cms:write, assets:write]
      parameters:
        - $ref: '#/components/parameters/ContentID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateImageUploadInput' }
      responses:
        '201':
          description: News cover upload target
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreatedUploadEnvelope' }
  /admin/content/news/{contentId}/assets/{assetId}:
    get:
      operationId: getNewsCoverStatus
      security: [{ bearerAuth: [] }]
      x-required-scopes: [cms:read]
      parameters:
        - $ref: '#/components/parameters/ContentID'
        - { name: assetId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: News cover scan and processing status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AssetStatusEnvelope' }
  /admin/content/news/{contentId}/assets/{assetId}/complete:
    post:
      operationId: completeNewsCoverUpload
      security: [{ bearerAuth: [] }]
      x-required-scopes: [cms:write, assets:write]
      parameters:
        - $ref: '#/components/parameters/ContentID'
        - { name: assetId, in: path, required: true, schema: { type: string } }
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CompleteImageUploadInput' }
      responses:
        '200': { $ref: '#/components/responses/ContentItem' }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    Locale: { name: locale, in: query, schema: { $ref: '#/components/schemas/Locale' } }
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    PageSize: { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
...
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 1 } }
    IfMatch: { name: If-Match, in: header, required: true, schema: { type: string, pattern: '^"[1-9][0-9]*"$' } }
    ContentModule: { name: module, in: path, required: true, schema: { $ref: '#/components/schemas/ContentModule' } }
    ContentID: { name: contentId, in: path, required: true, schema: { type: string, format: uuid } }
    Revision: { name: revision, in: path, required: true, schema: { type: integer, format: int64, minimum: 1 } }
  requestBodies:
...
    ContentWriteInput:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ContentWriteInput' }
  responses:
...
    Error:
      description: Standard API error envelope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    ContentItem:
      description: Typed content item
      headers:
        ETag: { schema: { type: string } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ContentItemEnvelope' }
    PublicContentList:
      description: Published localized content
      content:
        application/json:
          schema: { $ref: '#/components/schemas/PublicContentListEnvelope' }
  schemas:
    Locale:
      type: string
      enum: [zh-Hant, zh-Hans, en]
      default: zh-Hant
...
    ContentModule:
      type: string
      enum: [news, history, videos]
    ContentStatus:
      type: string
      enum: [draft, publishing, published, publish_failed, unpublishing, unpublish_failed, unpublished, archived]
    PageMeta:
      type: object
      required: [page, pageSize, total]
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer, format: int64 }
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
...
    AssetSummary:
      type: object
      required: [id]
      properties:
        id: { type: string }
    UploadTarget:
      type: object
      required: [url, method, headers, expiresAt]
      properties:
        url: { type: string, format: uri }
        method: { type: string }
        headers: { type: object, additionalProperties: { type: string } }
        expiresAt: { type: string, format: date-time }
    CreatedUpload:
      type: object
      required: [asset, uploadTarget]
      properties:
        asset: { $ref: '#/components/schemas/AssetSummary' }
        uploadTarget: { $ref: '#/components/schemas/UploadTarget' }
...
    CreateImageUploadInput:
      type: object
      additionalProperties: false
      required: [fileName, mimeType, sizeBytes]
      properties:
        fileName: { type: string, minLength: 1 }
        mimeType: { type: string, enum: [image/jpeg, image/png, image/webp] }
        sizeBytes: { type: integer, format: int64, minimum: 1, maximum: 10485760 }
    CompleteImageUploadInput:
      type: object
      additionalProperties: false
      required: [fileName, mimeType, sizeBytes, checksumSha256]
      properties:
        fileName: { type: string, minLength: 1 }
        mimeType: { type: string, enum: [image/jpeg, image/png, image/webp] }
        sizeBytes: { type: integer, format: int64, minimum: 1, maximum: 10485760 }
        checksumSha256: { type: string, pattern: '^[a-fA-F0-9]{64}$' }
    AssetStatus:
      type: object
      required: [id, namespace, ownerService, ownerType, ownerId, uploadStatus, scanStatus, processingStatus, visibility]
      properties:
        id: { type: string }
        namespace: { type: string }
        ownerService: { type: string }
        ownerType: { type: string }
        ownerId: { type: string }
        uploadStatus: { type: string, enum: [created, completed, failed] }
        scanStatus: { type: string, enum: [pending, clean, infected, failed] }
        processingStatus: { type: string, enum: [pending, ready, not_required, failed] }
        visibility: { type: string }
...
    ContentTranslation:
      type: object
      additionalProperties: false
      required: [locale, title]
      properties:
        locale: { $ref: '#/components/schemas/Locale' }
        title: { type: string, minLength: 1 }
        summary: { type: string }
        body: { type: string }
        dateLabel: { type: string }
        imageAlt: { type: string }
    ContentWriteInput:
      type: object
      additionalProperties: false
      required: [translations]
      properties:
        slug: { type: string }
        displayDate: { type: string, format: date }
        sortOrder: { type: integer, minimum: 1 }
        youtubeVideoId: { type: string, pattern: '^[A-Za-z0-9_-]{11}$' }
        coverAssetId: { type: string }
        featured: { type: boolean }
        homeEligible: { type: boolean }
        translations:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/ContentTranslation' }
    ContentItem:
      allOf:
        - $ref: '#/components/schemas/ContentWriteInput'
        - type: object
          required: [id, module, status, version, isPublished, createdBy, updatedBy, createdAt, updatedAt]
          properties:
            id: { type: string, format: uuid }
            module: { $ref: '#/components/schemas/ContentModule' }
            status: { $ref: '#/components/schemas/ContentStatus' }
            version: { type: integer, format: int64 }
            coverUrl: { type: string }
            isPublished: { type: boolean }
            publishedVersion: { type: integer, format: int64 }
            createdBy: { type: string }
            updatedBy: { type: string }
            publishedAt: { type: string, format: date-time }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
    ContentRevision:
      type: object
      required: [version, snapshot, createdBy, createdAt]
      properties:
        version: { type: integer, format: int64 }
        snapshot: { $ref: '#/components/schemas/ContentItem' }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
    PublicContentItem:
      type: object
      required: [id, title]
      properties:
        id: { type: string }
        title: { type: string }
        summary: { type: string }
        body: { type: string }
        dateLabel: { type: string }
        displayDate: { type: string, format: date }
        imageAlt: { type: string }
        imageUrl: { type: string }
        href: { type: string }
        youtubeVideoId: { type: string }
        sortOrder: { type: integer }
        featured: { type: boolean }
        homeEligible: { type: boolean }
...
    CreatedUploadEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/CreatedUpload' }
        meta: { type: object, additionalProperties: true }
        error: { type: ['null'] }
...
    ErrorEnvelope:
      type: object
      required: [data, meta, error]
      properties:
        data: { type: ['null'] }
        meta: { type: object, additionalProperties: true }
        error: { $ref: '#/components/schemas/ErrorBody' }
    ContentItemEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/ContentItem' }
        meta: { type: object, additionalProperties: true }
        error: { type: ['null'] }
    ContentListEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/ContentItem' } }
        meta: { $ref: '#/components/schemas/PageMeta' }
        error: { type: ['null'] }
    ContentRevisionListEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/ContentRevision' } }
        meta: { type: object, additionalProperties: true }
        error: { type: ['null'] }
    PublicContentListEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/PublicContentItem' } }
        meta: { type: object, additionalProperties: true }
        error: { type: ['null'] }
    HomeEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [news, videos]
          properties:
            news: { type: array, items: { $ref: '#/components/schemas/PublicContentItem' } }
            videos: { type: array, items: { $ref: '#/components/schemas/PublicContentItem' } }
        meta: { type: object, additionalProperties: true }
        error: { type: ['null'] }
    AssetStatusEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/AssetStatus' }
        meta: { type: object, additionalProperties: true }
        error: { type: ['null'] }